+++ src/components/BounceActivityContainer/Modals/ActivityDetailsModal/index.scss
